# core-banking-service/target 2/generated-sources/openapi/src/main/java/com/bca/core_banking_service/api/AccountsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bca.core_banking_service.api;

import com.bca.core_banking_service.dto.AccountCreate;
import com.bca.core_banking_service.dto.AccountResponse;
import com.bca.core_banking_service.dto.AmountRequest;
import com.bca.core_banking_service.dto.ErrorResponse;
import com.bca.core_banking_service.dto.TransactionResponse;
import com.bca.core_banking_service.dto.TransferRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-16T18:51:45.932800-05:00[America/Lima]")
@Validated
@Tag(name = "Accounts", description = "the Accounts API")
public interface AccountsApi {

    default AccountsApiDelegate getDelegate() {
        return new AccountsApiDelegate() {};
    }

    /**
     * POST /accounts/{accountId}/deposit : Depósito en cuenta
     *
     * @param accountId  (required)
     * @param amountRequest  (required)
     * @return Depósito exitoso (status code 200)
     */
    @Operation(
        operationId = "accountsAccountIdDepositPost",
        summary = "Depósito en cuenta",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Depósito exitoso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TransactionResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/accounts/{accountId}/deposit",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<TransactionResponse>> accountsAccountIdDepositPost(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId,
        @Parameter(name = "AmountRequest", description = "", required = true) @Valid @RequestBody Mono<AmountRequest> amountRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().accountsAccountIdDepositPost(accountId, amountRequest, exchange);
    }


    /**
     * POST /accounts/{accountId}/withdraw : Retiro de cuenta
     *
     * @param accountId  (required)
     * @param amountRequest  (required)
     * @return Retiro exitoso (status code 200)
     *         or Solicitud inválida (status code 400)
     *         or Saldo insuficiente (status code 409)
     */
    @Operation(
        operationId = "accountsAccountIdWithdrawPost",
        summary = "Retiro de cuenta",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Retiro exitoso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TransactionResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Saldo insuficiente")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/accounts/{accountId}/withdraw",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<TransactionResponse>> accountsAccountIdWithdrawPost(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId,
        @Parameter(name = "AmountRequest", description = "", required = true) @Valid @RequestBody Mono<AmountRequest> amountRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().accountsAccountIdWithdrawPost(accountId, amountRequest, exchange);
    }


    /**
     * POST /accounts : Crear cuenta bancaria
     *
     * @param accountCreate  (required)
     * @return Cuenta creada (status code 201)
     *         or Solicitud inválida (status code 400)
     */
    @Operation(
        operationId = "accountsPost",
        summary = "Crear cuenta bancaria",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Cuenta creada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AccountResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/accounts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<AccountResponse>> accountsPost(
        @Parameter(name = "AccountCreate", description = "", required = true) @Valid @RequestBody Mono<AccountCreate> accountCreate,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().accountsPost(accountCreate, exchange);
    }


    /**
     * POST /accounts/transfer : Transferencia entre cuentas
     *
     * @param transferRequest  (required)
     * @return Transferencia exitosa (status code 200)
     */
    @Operation(
        operationId = "accountsTransferPost",
        summary = "Transferencia entre cuentas",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Transferencia exitosa", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TransactionResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/accounts/transfer",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<TransactionResponse>> accountsTransferPost(
        @Parameter(name = "TransferRequest", description = "", required = true) @Valid @RequestBody Mono<TransferRequest> transferRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().accountsTransferPost(transferRequest, exchange);
    }

}