# core-banking-service/target 2/generated-sources/openapi/src/main/java/com/bca/core_banking_service/api/CreditsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bca.core_banking_service.api;

import com.bca.core_banking_service.dto.AmountRequest;
import com.bca.core_banking_service.dto.CreditCreate;
import com.bca.core_banking_service.dto.CreditPaymentResponse;
import com.bca.core_banking_service.dto.CreditResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-16T18:51:45.932800-05:00[America/Lima]")
@Validated
@Tag(name = "Credits", description = "the Credits API")
public interface CreditsApi {

    default CreditsApiDelegate getDelegate() {
        return new CreditsApiDelegate() {};
    }

    /**
     * POST /credits/{creditId}/payment : Pago de crédito
     *
     * @param creditId  (required)
     * @param amountRequest  (required)
     * @return Pago exitoso (status code 200)
     *         or Monto inválido (status code 409)
     */
    @Operation(
        operationId = "creditsCreditIdPaymentPost",
        summary = "Pago de crédito",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Pago exitoso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreditPaymentResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Monto inválido")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/credits/{creditId}/payment",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CreditPaymentResponse>> creditsCreditIdPaymentPost(
        @Parameter(name = "creditId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("creditId") String creditId,
        @Parameter(name = "AmountRequest", description = "", required = true) @Valid @RequestBody Mono<AmountRequest> amountRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().creditsCreditIdPaymentPost(creditId, amountRequest, exchange);
    }


    /**
     * POST /credits : Crear crédito
     *
     * @param creditCreate  (required)
     * @return Crédito creado (status code 201)
     *         or Crédito no aprobado (status code 400)
     */
    @Operation(
        operationId = "creditsPost",
        summary = "Crear crédito",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Crédito creado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreditResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Crédito no aprobado")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/credits",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CreditResponse>> creditsPost(
        @Parameter(name = "CreditCreate", description = "", required = true) @Valid @RequestBody Mono<CreditCreate> creditCreate,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().creditsPost(creditCreate, exchange);
    }

}